+++ Examples/Resources/WaitForADDomain/7-WaitForADDomain_WaitForDomainControllerIgnoringAuthenticationErrors_Config.ps1
@@ -1,6 +1,6 @@
 <#PSScriptInfo
 .VERSION 1.0
-.GUID 2ada2ead-8736-4d5e-9587-e14bacc28761
+.GUID 6b60ca02-019b-481a-ac34-a2f24df09ffd
 .AUTHOR Microsoft Corporation
 .COMPANYNAME Microsoft Corporation
 .COPYRIGHT (c) Microsoft Corporation. All rights reserved.
@@ -20,15 +20,10 @@
 <#
     .DESCRIPTION
         This configuration will wait for an Active Directory domain controller
-        to respond within 300 seconds (default) in the domain 'contoso.com'
-        before returning and allowing the configuration to continue to run.
-        If the timeout is reached the node will be restarted up to two times
-        and again wait after each restart. If no domain controller is found
-        after the second restart an error will be thrown.
-        This will use the user credential passed in the built-in PsDscRunAsCredential
-        parameter when determining if the domain is available.
+        to respond within the default period, and ignore any authentication
+        errors.
 #>
-Configuration WaitForADDomain_WaitForDomainControllerWithReboot_Config
+Configuration WaitForADDomain_WaitForDomainControllerIgnoringAuthenticationErrors_Config
 {
     param
     (
@@ -44,10 +39,10 @@ Configuration WaitForADDomain_WaitForDomainControllerWithReboot_Config
     {
         WaitForADDomain 'contoso.com'
         {
-            DomainName   = 'contoso.com'
-            RestartCount = 2
+            DomainName             = 'contoso.com'
+            WaitForValidCredentials = $true
 
-            PsDscRunAsCredential = $Credential
+            PsDscRunAsCredential   = $Credential
         }
     }
 }